+++ packages/terra-worklist-data-grid/src/subcomponents/Row.jsx
@@ -70,11 +70,6 @@ const propTypes = {
    * A zero-based index indicating which column represents the row header.
    */
   rowHeaderIndex: PropTypes.number,
-
-  /**
-   * Id of the column in the row that is selected.
-   */
-  selectedCellColumnId: PropTypes.string,
 };
 
 const defaultProps = {
@@ -96,22 +91,21 @@ function Row(props) {
     rowHeaderIndex,
     onRowSelect,
     onCellSelect,
-    selectedCellColumnId,
   } = props;
 
   const theme = useContext(ThemeContext);
 
   const columnIndexOffSet = hasRowSelection ? 1 : 0;
 
-  const handleCellSelect = useCallback((rowIdColId, coordinates, cellSelectable = true) => {
-    if (hasRowSelection) {
+  const handleCellSelect = useCallback((selectionDetails) => {
+    if (hasRowSelection || selectionDetails.multiSelect) {
       if (onRowSelect) {
-        onRowSelect(rowIdColId.rowId, rowIndex, coordinates);
+        onRowSelect(selectionDetails);
       }
     } else if (onCellSelect) {
-      onCellSelect(rowIdColId, coordinates, cellSelectable);
+      onCellSelect(selectionDetails);
     }
-  }, [hasRowSelection, onCellSelect, onRowSelect, rowIndex]);
+  }, [hasRowSelection, onCellSelect, onRowSelect]);
 
   const getCellData = (cellRowIndex, cellColumnIndex, cellData, rowId) => {
     const columnId = displayedColumns[cellColumnIndex].id;
@@ -124,7 +118,7 @@ function Row(props) {
         rowIndex={cellRowIndex}
         columnIndex={cellColumnIndex}
         key={`${rowId}_${columnId}`}
-        isSelected={!hasRowSelection && selectedCellColumnId === columnId}
+        isSelected={!hasRowSelection && cellData.isSelected}
         isMasked={cellData.isMasked}
         isSelectable={cellData.isSelectable}
         isRowHeader={isRowHeader}
@@ -150,7 +144,10 @@ function Row(props) {
 
   return (
     <tr
-      className={cx([isSelected ? 'row-selected' : 'worklist-data-grid-row', theme.className])}
+      className={cx('worklist-data-grid-row', {
+        selected: isSelected,
+        selectable: hasRowSelection,
+      }, theme.className)}
       // eslint-disable-next-line react/forbid-dom-props
       style={{ height }}
     >